Move checks into a new folder
