+++ QModManager/Checks/NitroxCheck.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
 
-namespace QModManager
+namespace QModManager.Checks
 {
     internal static class NitroxCheck
     {
