+++ QModManager/QModManager.csproj
@@ -120,11 +120,11 @@
     <Compile Include="Obsolete\PatchManager.cs" />
     <Compile Include="API\Messages.cs" />
     <Compile Include="API\QModAPI.cs" />
-    <Compile Include="NitroxCheck.cs" />
+    <Compile Include="Checks\NitroxCheck.cs" />
     <Compile Include="Utility\IOUtilities.cs" />
     <Compile Include="Utility\Logger.cs" />
     <Compile Include="OptionsManager.cs" />
-    <Compile Include="PirateCheck.cs" />
+    <Compile Include="Checks\PirateCheck.cs" />
     <Compile Include="Prefab Debugger\HierarchyItem.cs" />
     <Compile Include="Prefab Debugger\LogMessage.cs" />
     <Compile Include="Prefab Debugger\PrefabDebugger.cs" />
@@ -136,7 +136,7 @@
     <Compile Include="Utility\NetworkUtilities.cs" />
     <Compile Include="Utility\PlayerPrefsExtra.cs" />
     <Compile Include="Utility\ReflectionHelper.cs" />
-    <Compile Include="VersionCheck.cs" />
+    <Compile Include="Checks\VersionCheck.cs" />
   </ItemGroup>
   <Import Project="API\SMLHelper\SMLHelper.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
